refactor: 범용성에 따른 유틸 코드 분리
- 유틸 패키지 생성 후 활용도에 따른 분류
