+++ src/main/java/common/constant/RegexConstants.java
@@ -1,4 +1,4 @@
-package constant;
+package common.constant;
 
 public class RegexConstants {
     public static final String NUMBER_ONLY_REGEX = "^[0-9]+$"; // 숫자 판별
