+++ src/main/java/validator/Validator.java
@@ -1,5 +1,14 @@
 package validator;
 
+import static common.constant.NumberConstants.LOTTO_NUMBER_END;
+import static common.constant.NumberConstants.LOTTO_NUMBER_START;
+import static common.constant.RegexConstants.NUMBER_ONLY_REGEX;
+import static common.utils.ValidationUtils.checkEmptyInput;
+import static common.utils.ValidationUtils.checkInvalidForm;
+import static common.utils.ValidationUtils.checkNullInput;
+import static common.utils.ValidationUtils.checkOutOfRange;
+import static common.utils.ValidationUtils.checkWhitespaceOnlyInput;
+
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -12,36 +21,15 @@ public class Validator {
         checkWhitespaceOnlyInput(input, ErrorMessages.WHITESPACE_ONLY_INPUT.getMessage());
     }
 
-    public static void checkNullInput(String input, String errorMessage) {
-        if (input == null) {
-            throw new IllegalArgumentException(errorMessage);
-        }
+    // 숫자 여부 유효성 검사
+    public static void validateNumber(String input) {
+        checkInvalidForm(input, NUMBER_ONLY_REGEX, ErrorMessages.NOT_NUMBER.getMessage());
     }
 
-    public static void checkEmptyInput(String input, String errorMessage) {
-        if (input.isEmpty()) {
-            throw new IllegalArgumentException(errorMessage);
-        }
-    }
-
-    public static void checkWhitespaceOnlyInput(String input, String errorMessage) {
-        if (input.trim().isEmpty()) {
-            throw new IllegalArgumentException(errorMessage);
-        }
-    }
-
-    // 주어진 정규표현식에 따른 유효성 검사
-    public static void validateInvalidForm(String input, String regexPattern, String errorMessage) {
-        if (!input.matches(regexPattern)) {
-            throw new IllegalArgumentException(errorMessage);
-        }
-    }
-
-    // 주어진 범위에 따른 유효성 검사 (이상 값, 이하 값)
-    public static void validateOutOfRange(int number, int rangeStart, int rangeEnd, String errorMessage) {
-        if (number < rangeStart || number > rangeEnd) {
-            throw new IllegalArgumentException(errorMessage);
-        }
+    // 로또 범위 유효성 검사
+    public static void validateLottoNumberRange(int number) {
+        checkOutOfRange(number, LOTTO_NUMBER_START, LOTTO_NUMBER_END,
+                ErrorMessages.LOTTO_NUMBER_OUT_OF_RANGE.getMessage());
     }
 
     // set의 특성을 활용한 리스트 내의 중복된 요소 판별
